+++ demos/cli/src/main.rs
@@ -178,13 +178,14 @@ fn run3d<F: fidget::eval::Function + fidget::render::RenderHints>(
     };
     let shape = shape.apply_transform(mat.into());
 
-    let mut depth = vec![];
-    let mut color = vec![];
+    let mut depth = Default::default();
+    let mut norm = Default::default();
     for _ in 0..settings.n {
-        (depth, color) = cfg.run(shape.clone()).unwrap();
+        (depth, norm) = cfg.run(shape.clone()).unwrap();
     }
 
     let out = if mode_color {
+        let color = norm.to_color();
         depth
             .into_iter()
             .zip(color)
@@ -269,7 +270,7 @@ fn run2d<F: fidget::eval::Function + fidget::render::RenderHints>(
             ..Default::default()
         };
         if sdf {
-            let mut image = vec![];
+            let mut image = fidget::render::Image::default();
             for _ in 0..settings.n {
                 image = cfg
                     .run::<_, fidget::render::SdfRenderMode>(shape.clone())
@@ -280,7 +281,7 @@ fn run2d<F: fidget::eval::Function + fidget::render::RenderHints>(
                 .flat_map(|a| [a[0], a[1], a[2], 255].into_iter())
                 .collect()
         } else {
-            let mut image = vec![];
+            let mut image = fidget::render::Image::default();
             for _ in 0..settings.n {
                 image = cfg
                     .run::<_, fidget::render::DebugRenderMode>(shape.clone())